Change unorded map to array to increase speed of code significantly

+++ resources/message_system/src/message_distributer.cpp
@@ -12,7 +12,7 @@ MessageDistributer::~MessageDistributer()
 
 void MessageDistributer::publish(Topics topic, const BaseMsg& msg)
 {
-    for (auto& subscriber_callback : subscriber_callbacks_map[topic])
+    for (auto& subscriber_callback : subscriber_callbacks_map[static_cast<std::size_t>(topic)])
     {
         subscriber_callback->operator()(msg);
     }
@@ -20,12 +20,12 @@ void MessageDistributer::publish(Topics topic, const BaseMsg& msg)
 
 void MessageDistributer::add_subscriber(Topics topic, std::shared_ptr<std::function<void(const BaseMsg& msg)>> subscriber_callback)
 {
-    subscriber_callbacks_map[topic].push_back(subscriber_callback);
+    subscriber_callbacks_map[static_cast<std::size_t>(topic)].push_back(subscriber_callback);
 }
 
 void MessageDistributer::remove_subscriber(Topics topic, std::shared_ptr<std::function<void(const BaseMsg& msg)>> subscriber_callback)
 {
-    auto& topic_callbacks = subscriber_callbacks_map[topic];
+    auto& topic_callbacks = subscriber_callbacks_map[static_cast<std::size_t>(topic)];
     auto itterator = std::find_if(topic_callbacks.begin(), topic_callbacks.end(), 
         [&](const std::shared_ptr<std::function<void(const BaseMsg&)>>& callback) 
         {
@@ -40,37 +40,37 @@ void MessageDistributer::remove_subscriber(Topics topic, std::shared_ptr<std::fu
 
 void MessageDistributer::add_server(Service service, std::shared_ptr<std::function<void(const BaseService::Request&, BaseService::Response&)>> server_callback)
 {
-    if (server_callbacks_map[service] != nullptr)
+    if (server_callbacks_map[static_cast<std::size_t>(service)] != nullptr)
     {
         std::cout << "Error: Server already exists for service" << std::endl;
         return;
     }
-    server_callbacks_map[service] = server_callback;
+    server_callbacks_map[static_cast<std::size_t>(service)] = server_callback;
 }
 
 void MessageDistributer::remove_server(Service service, std::shared_ptr<std::function<void(const BaseService::Request&, BaseService::Response&)>> server_callback)
 {
-    if(server_callbacks_map[service] == nullptr)
+    if(server_callbacks_map[static_cast<std::size_t>(service)] == nullptr)
     {
         std::cout << "Error: Server does not exist for service" << std::endl;
         return;
     }
-    if (server_callbacks_map[service] != server_callback)
+    if (server_callbacks_map[static_cast<std::size_t>(service)] != server_callback)
     {
         std::cout << "Error: Cannot remove callback as it is incorrect" << std::endl;
         return;
     }
-    server_callbacks_map[service] = nullptr;
+    server_callbacks_map[static_cast<std::size_t>(service)] = nullptr;
 }
 
 void MessageDistributer::send_request(Service service, const BaseService::Request& request, BaseService::Response& response, std::shared_ptr<std::function<void(const BaseService::Response&)>> client_callback)
 {
-    if (server_callbacks_map[service] == nullptr)
+    if (server_callbacks_map[static_cast<std::size_t>(service)] == nullptr)
     {
         std::cout << "Error: Server does not exist for service" << std::endl;
         return;
     }
 
-    server_callbacks_map[service]->operator()(request, response);
+    server_callbacks_map[static_cast<std::size_t>(service)]->operator()(request, response);
     client_callback->operator()(response);
 }
